splits repl.ts into multiple files
