+++ src/juliaDebug.ts
@@ -12,7 +12,7 @@ import { Subject } from 'await-notify';
 import * as readline from 'readline';
 import { generatePipeName } from './utils';
 import { uuid } from 'uuidv4';
-import { replStartDebugger } from './repl';
+import { replStartDebugger } from './interactive/repl';
 import * as vscode from 'vscode';
 import { getCrashReportingPipename } from './telemetry';
